refactor: proper name for event card component
