+++ dashboard/src/components/EventCard.vue
