+++ dashboard/src/pages/Home.vue
@@ -2,16 +2,16 @@
   <Header></Header>
   <div class=" bg-gray-50 justify-center align-middle flex">
     <div class="flex container py-8 px-4 h-screen">
-      <main class="w-full">
-        <div class="greet-user">
+      <main class="w-full space-y-4">
+        <div class="greet-user" v-if="profile.data">
           <div class="text-3xl font-semibold my-2">
-            Hello <span v-if="profile.data">{{ profile.data.full_name.split(" ")[0] }}</span>,
+            Hello {{ profile.data.full_name.split(" ")[0] }},
           </div>
           <div class=" text-gray-600">
             This is the organizer's dashboard. You can manage your chapters, events, and more from here.
           </div>
         </div>
-        <div class="py-4">
+        <div>
           <div class="text-lg font-semibold my-2">Your Chapters</div>
           <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
             <div
@@ -23,14 +23,14 @@
             </div>
           </div>
         </div>
-        <div class="py-4">
+        <div>
           <div class="text-lg font-semibold my-2">Scheduled Events</div>
           <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
             <div
               v-for="event in scheduled_events.data"
               :key="event.name"
             >
-              <EventCardDashboard :event="event"></EventCardDashboard>
+              <EventCard :event="event" />
             </div>
           </div>
         </div>
@@ -40,12 +40,12 @@
 </template>
 
 <script setup>
-import { createResource, createListResource, usePageMeta, ListView, Badge } from 'frappe-ui'
-import Header from '@/components/Header.vue'
-import ChapterCard from '@/components/ChapterCard.vue'
+import { watch } from 'vue'
 import { session } from '@/data/session.js'
-import { watch, ref } from 'vue'
-import EventCardDashboard from '@/components/EventCardDashboard.vue'
+import Header from '@/components/Header.vue'
+import EventCard from '@/components/EventCard.vue'
+import ChapterCard from '@/components/ChapterCard.vue'
+import { createResource, createListResource, usePageMeta } from 'frappe-ui'
 
 let chapter = createListResource({
   doctype: 'FOSS Chapter',
